+++ internal/config/secrets.go
@@ -13,8 +13,11 @@ import (
 const secretsFileName = "secrets.json"
 
 type Secrets struct {
-	DbPassword      string `json:"dbPassword"`
-	AdminUiPassword string `json:"adminUiPassword,omitempty"`
+	DbPassword           string `json:"dbPassword"`
+	AdminUiPassword      string `json:"adminUiPassword,omitempty"`
+	AiLabScsPassword     string `json:"aiLabScsPassword,omitempty"`
+	AiLabJupyterPassword string `json:"aiLabJupyterPassword,omitempty"`
+	AiLabBfsPassword     string `json:"aiLabBfsPassword,omitempty"`
 }
 
 func SecretsFilePath(deployment DeploymentDir) (string, error) {